% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb.R
\name{msc_plumb}
\alias{msc_plumb}
\title{Runs the MultiSC-Daemon API}
\usage{
msc_plumb(port = plumber::get_option_or_env("plumber.port", 8000), ...)
}
\arguments{
\item{port}{Port to run the API on. Default: 8000}

\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Launches the Plumber API for the MultiSC-Daemon API.
}
