% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{write_genes}
\alias{write_genes}
\title{Write genes to JSON file}
\usage{
write_genes(rds = readRDS(file = "data.rds"), path = "genes.json")
}
\arguments{
\item{rds}{A Seurat object. Default reads from "data.rds".}

\item{path}{The file path to write the JSON file to. Default is "genes.json".}
}
\value{
A character vector of unique gene names.
}
\description{
Writes the gene names from a Seurat object to a JSON file.
}
